separate classes from main
